requirements.txt
langgraph
litellm
selenium
webdriver_manager
selenium-stealth
colorama
python-dotenv
